W239: the nd21-100 band — +4 MATCH (6598, 80.2%), nd -1163
0 MATCHes lost, 0 new over-window, one window cleared. **First time past 80.2%.**

The wave targeted the 116 non-floor functions at nd21-100 — the closest population not already written
off as compiler floor — where a residual of five to twenty-five instructions can be read directly and
attributed to a specific source defect. Only 22 had a matched template sibling, so most were worked
straight off the function diff.

**MATCHes, and each names a distinct idiom worth keeping:**

- `fclShopMisc 003f7fe0` nd93 -> 0. Our source masked a colour parameter at **six** later draw calls;
  retail masks it only for the initial `FUN_0040e3c0` call and passes the original value to all six.
  **Deleting six invented masks reproduced retail exactly.**
- `fclShopMisc 003f03e0` nd100 -> 0. `socialLink` declared `s16` where the original used `int`, so
  retail's constants live directly in `$s0`; narrowing explicitly at only the one call that needs it
  reproduces retail's `dsll32`/`dsra32` at that boundary.
- `mt_evtMain 00366eb0` nd95 -> 0. **The early-return guard.** Our tail was
  `if (node != 0) { stores; call; } return 0;`; retail uses `if (node == 0) return 0;` followed by an
  unconditional body. Semantically identical, different branch shape — b210 follows the source's shape
  closely, so when the residual is branch orientation or a join placement, rewrite the control flow
  rather than the types.
- `btlBoss 002fcaa0` nd26 -> 0. Direct early returns replaced by result-local assignments and structured
  breaks joining at one epilogue — the idiom used by `FUN_002fcf50`, the already-matched function
  **immediately below it** in the same file.

**The clearest strategic finding: yield tracks how little a file has been worked, not how low the diff
is.** J239Evt's first pass over the event files kept nothing from fifteen targets — every residual
classified against the real prologue as call-argument setup order, preheader constant placement, or
cyclic colouring with matching save counts. Those files have been mined in six previous waves. Meanwhile
the barely-touched files gave up all four matches. I broadcast that mid-wave and J239Evt then returned to
find `00366eb0`, so the correction paid immediately.

**The excess-saved-register lens works on the integer side too**, extending wave 238's float result.
`mdlFile 003252a0` nd402 -> 210 and `00325500` nd442 -> 270, both reaching retail's exact save counts
(3 GPRs -> 1, and 1 GPR/2 FPR), via 32-bit pointer types, `LQ`/`SQ` quadword aggregate copies, and
callback-record fields instead of separate symbol bases. `00320080` cleared its window (464/432 ->
432/432, nd249 -> 168) once three separate symbol bases were recognised as **one 0x10-byte dispatch
record** with create/callback at +4/+8/+0xc and an erroneous second function-pointer dereference removed.

Near-misses now parked one or two instructions out: `fclShopMisc 003fee40` **nd94 -> 2** (a recovered
0x1c `ShopRequirements` record with signed `s16` fields; residual is `slti $v0` vs our `slti $at`),
`frFontRaster 003b3f90` **nd29 -> 1** at exactly 528/528 (a `u8` nibble with an explicit 0xf0 mask
restoring retail's `ANDI 0xf` / `SLL` / `ANDI 0xf0` / `ANDI 0xff` sequence; residual is one commutative
`addu`). Also `k_event 001c8120` nd846 -> 609 from a substantial retail-backed reconstruction, and
`itfMesManager 003a53b0` nd22 -> 8 via typed `u8` call aliases forcing retail's `ANDI` at each boundary.

**My framescan over-reports and agents caught it four times.** At `k_fldFrame 001ae580`,
`k_footstep 001da020` and two `_h_camp_status` functions the real prologues agree on saved registers — in
the Camp cases *retail* saves more than we do, so the correct reading is a missing aggregate, not excess
GPR lifetime. That correction is worth more than the rows it invalidated: it repoints `_h_camp_status`
at a deficit-reconstruction lens.

Bands after this wave: 429 at nd<=20, 1974 at nd21-100, 2710 at nd101-500, 1174 above.

+++ src/Yajima/y_misc.c
@@ -34,7 +34,7 @@ char cGpffffb9d8;
 char cGpffffb9ec;
 int iGpffffb5c0;
 int *piGpffffa850;
-u32 uGpffffb9ec;
+u8 uGpffffb9ec;
 u32 uGpffffb9f4;
 extern u8 cGpffffb9c0;
 extern u8* iGpffffb9b8;
